requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
colorama==0.4.6
frozenlist==1.7.0
fuzzywuzzy==0.18.0
greenlet==3.2.3
idna==3.10
multidict==6.6.3
numpy==2.3.1
packaging==25.0
pillow==11.3.0
playwright==1.53.0
propcache==0.3.2
pyee==13.0.0
pytesseract==0.3.13
requests==2.32.4
scipy==1.16.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
